flake8 updates for tests/perf

+++ tests/perf/write/create_empty.py
@@ -28,23 +28,31 @@ if iter_type not in ("row", "col", "block"):
     raise ValueError("invalid iteration type")
 
 if filepath.startswith("hdf5://"):
-    f = h5pyd.File(filepath, 'w')
+    f = h5pyd.File(filepath, "w")
 else:
-    f = h5py.File(filepath, 'w')
+    f = h5py.File(filepath, "w")
 
-chunks = [500, 1000] # 3.8 MB/chunk
+chunks = [500, 1000]  # 3.8 MB/chunk
 if chunks[0] > nrows:
     chunks[0] = nrows
 if chunks[1] > ncols:
     chunks[1] = ncols
 chunks = tuple(chunks)
-dset = f.create_dataset("dset", (nrows, ncols), dtype='f8', chunks=chunks)
+dset = f.create_dataset("dset", (nrows, ncols), dtype="f8", chunks=chunks)
 print("dset:", dset)
 print("dset chunks:", dset.chunks)
 print("dset id:", dset.id.id)
 
-dt = [("start", "f8"), ("done", "f8"), ('status', "i4"), 
-       ("x", "i4"), ("y", "i4"), ("nrow", "i4"), ("ncol", "i4"), ('pod', "S40")]
+dt = [
+    ("start", "f8"),
+    ("done", "f8"),
+    ("status", "i4"),
+    ("x", "i4"),
+    ("y", "i4"),
+    ("nrow", "i4"),
+    ("ncol", "i4"),
+    ("pod", "S40"),
+]
 
 num_tasks = 0
 if iter_type == "block":
@@ -61,8 +69,8 @@ if num_tasks < 1000:
 else:
     task_chunks = (1000,)
 
-arr = np.zeros((num_tasks,), dtype=dt)  
-index = 0 
+arr = np.zeros((num_tasks,), dtype=dt)
+index = 0
 
 if iter_type == "block":
     for s in dset.iter_chunks():
@@ -70,16 +78,16 @@ if iter_type == "block":
         y = s[1].start
         nrow = s[0].stop - s[0].start
         ncol = s[1].stop - s[1].start
-        arr[index] = (0,0,0,x,y,nrow,ncol,"")
+        arr[index] = (0, 0, 0, x, y, nrow, ncol, "")
         index += 1
-        
+
 elif iter_type == "row":
     for i in range(nrows):
         x = i
         y = 0
         nrow = 1
         ncol = ncols
-        arr[index] = (0,0,0,x,y,nrow,ncol,"")
+        arr[index] = (0, 0, 0, x, y, nrow, ncol, "")
         index += 1
 else:
     # col
@@ -88,7 +96,7 @@ else:
         y = i
         nrow = nrows
         ncol = 1
-        arr[index] = (0,0,0,x,y,nrow,ncol,"")
+        arr[index] = (0, 0, 0, x, y, nrow, ncol, "")
         index += 1
 
 if index != num_tasks:
@@ -100,4 +108,3 @@ f.create_dataset("task_list", data=arr, chunks=task_chunks)
 
 f.close()
 print("created chunk_list dataset")
-
